+++ misc/load_waypoints/launch/load_waypoints.launch
@@ -1,3 +1,3 @@
 <launch>
-    <node pkg="load_waypoints" type="load_waypoints_server.py" name="load_waypoints_server" output="screen" ns="caffeine"/>
+    <node pkg="load_waypoints" type="navigate_waypoints.py" name="load_waypoints_server" output="screen" ns="caffeine"/>
 </launch>
