+++ misc/load_waypoints/scripts/static_waypoints.json
@@ -4,25 +4,37 @@
             "id" : 0,
             "longitude" : 50.0,
             "latitude" : 90.0,
-            "description"  : "Entrance of No-Man's Land"
+            "description"  : "Entrance of No-Man's Land",
+            "frame_id" : "caffeine/odom",
+            "pose_test" : [11.2,0,0,0,0,0],
+            "pose_test_norm" : [10,0,0,0,0,0]
         },
         {
             "id" : 1,
             "longitude" : 11.0,
             "latitude" : 22.0,
-            "description"  : "Exit of No-Man's Land"
+            "description"  : "Exit of No-Man's Land",
+            "frame_id" : "caffeine/odom",
+            "pose_test" : [15.7,-4.06,0,0,0,-0.547],
+            "pose_test_norm" : [16.2, -4.46, 0, 0, 0, -0.967]
         },
         {
             "id" : 2,
             "longitude" : 12.0,
             "latitude" : 34.0,
-            "description"  : "Inside of No-Man's Land"
+            "description"  : "Inside of No-Man's Land",
+            "frame_id" : "caffeine/odom",
+            "pose_test" : [18.95,-9.56,0,0,0,-2.187],
+            "pose_test_norm" : [18.028, -10.42, 0, 0, 0, -2.185676849181859]
         },
         {
             "id" : 3,
             "longitude" : 54.0,
             "latitude" : 100.0,
-            "description"  : "Inside of No-Man's Land"
+            "description"  : "Inside of No-Man's Land",
+            "frame_id" : "caffeine/odom",
+            "pose_test" : [14.56,-14.39,0,0,0,-1.99],
+            "pose_test_norm" : [15.115, -14.65, 0, 0, 0, -1.9989176939225397]
         }
     ]   
 }
