--- nav/nav_goal/package.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>nav_goal</name>
-  <version>0.0.0</version>
-  <description>The nav_goal package allows publishing to the move_base/goal topic to tell the robot where to move </description>
-
-  <maintainer email="yviel@todo.todo">yviel</maintainer>
-
-  <license>TODO</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <build_depend>actionlib</build_depend>
-  <build_depend>move_base_msgs</build_depend>
-  <build_depend>rospy</build_depend>
-  <build_depend>tf</build_depend>
-
-  <build_export_depend>actionlib</build_export_depend>
-  <build_export_depend>move_base_msgs</build_export_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <build_export_depend>tf</build_export_depend>
-
-  <exec_depend>actionlib</exec_depend>
-  <exec_depend>move_base_msgs</exec_depend>
-  <exec_depend>rospy</exec_depend>
-  <exec_depend>tf</exec_depend>
-
-</package>
-
