+++ nav/nav_stack/config/base_local_planner_params.yaml
@@ -3,7 +3,7 @@ NavfnROS:
   # Specifies whether or not to allow navfn to create plans that traverse unknown space.
   allow_unknown: true
   # A tolerance on the goal point for the planner.
-  default_tolerance: 0.1
+  default_tolerance: 0.3
 
 # Plans robot movement
 TrajectoryPlannerROS:
