+++ nav/nav_stack/launch/move_base.launch
@@ -2,7 +2,7 @@
 
     <node pkg="move_base" type="move_base" respawn="false" name="move_base" output="screen">
         <remap from="scan" to="caffeine/scan"/>
-        <remap from="odom" to="caffeine/odometry/loca"/>
+        <remap from="odom" to="caffeine/odometry/local"/>
         <remap from="cmd_vel" to="caffeine/nav_vel"/>
 
         <!-- Load common parameters for both the global and local costmaps -->
@@ -16,5 +16,7 @@
         <!-- Load navigation parameters -->
         <rosparam file="$(find nav_stack)/config/base_local_planner_params.yaml" command="load" />
     </node>
+
+    <node name="teleop_twist_keyboard" pkg="teleop_twist_keyboard" type="teleop_twist_keyboard.py"/>
 
 </launch>
